remove db sync and allow db migration plugin to run migrations
