+++ rundeckapp/grails-app/init/rundeckapp/cli/RundeckCliOptions.groovy
@@ -26,7 +26,6 @@ class RundeckCliOptions {
     boolean debug = false
     boolean rewrite = false
     boolean useJaas = true
-    boolean dbsync = false
     boolean rollback = false
 
     String baseDir
@@ -37,6 +36,5 @@ class RundeckCliOptions {
     String dataDir
     String projectDir
     String logDir
-    String migrationsDir
     String tag
 }
