+++ rundeckapp/templates/config/rundeck-config.properties.template
@@ -12,8 +12,12 @@ dataSource.url = jdbc:h2:file:${server.datastore.path};DB_CLOSE_ON_EXIT=FALSE
 
 dataSource.dbCreate = none
 dataSource.url = jdbc:h2:file:${server.datastore.path};MVCC=true
+<<<<<<< HEAD
 grails.plugin.migration.updateOnStart=true
 >>>>>>> Convert rundeck core application to use grails db-migration for database changes
+=======
+grails.plugin.databasemigration.updateOnStart=true
+>>>>>>> remove db sync and allow db migration plugin to run migrations
 # Pre Auth mode settings
 rundeck.security.authorization.preauthenticated.enabled=false
 rundeck.security.authorization.preauthenticated.attributeName=REMOTE_USER_GROUPS
